refactor(scripts/build-cache): Use only one script to build cache and hash everything without upload
